Move SemanticScholarFetcher to a more appropriate package
