+++ src/main/java/org/jabref/gui/entryeditor/citationrelationtab/RelatedEntriesFetcher.java
@@ -1,4 +1,4 @@
-package org.jabref.gui.entryeditor.citationrelationtab.semanticscholar;
+package org.jabref.gui.entryeditor.citationrelationtab;
 
 import java.util.List;
