+++ src/main/java/org/jabref/logic/importer/citationrelation/AuthorResponse.java
@@ -1,4 +1,4 @@
-package org.jabref.gui.entryeditor.citationrelationtab.semanticscholar;
+package org.jabref.logic.importer.citationrelation;
 
 public class AuthorResponse {
     private String authorId;
