+++ src/main/java/org/jabref/logic/importer/citationrelation/PaperDetails.java
@@ -1,4 +1,4 @@
-package org.jabref.gui.entryeditor.citationrelationtab.semanticscholar;
+package org.jabref.logic.importer.citationrelation;
 
 import java.util.List;
 import java.util.stream.Collectors;
