+++ src/main/java/org/jabref/logic/importer/citationrelation/ReferenceDataItem.java
@@ -1,4 +1,4 @@
-package org.jabref.gui.entryeditor.citationrelationtab.semanticscholar;
+package org.jabref.logic.importer.citationrelation;
 
 public class ReferenceDataItem {
     private PaperDetails citedPaper;
